requirements.txt
langchain
langgraph
langsmith
openai
tavily-python
duckdb
pandas
boto3
pydantic
python-dotenv
pypdf
langchain-openai
fastapi
uvicorn
python-multipart